requirements.txt
streamlit
groq
pdfminer.six
python-docx
docx2txt
pytesseract
pillow